LayoutTests/http/wpt/web-animations/interfaces/AnimationEffectTiming/direction-expected.txt


FAIL Test default value assert_equals: expected (string) "normal" but got (undefined) undefined
PASS set direction to a valid keyword 
